Feature/protegendo rotas

+++ src/main/java/com/hav/imobiliaria/service/UsuarioService.java
@@ -7,12 +7,14 @@ import com.hav.imobiliaria.controller.mapper.imovel.ImovelGetMapper;
 import com.hav.imobiliaria.controller.mapper.usuario.UsuarioGetMapper;
 import com.hav.imobiliaria.controller.mapper.usuario.UsuarioPostMapper;
 import com.hav.imobiliaria.controller.mapper.usuario.UsuarioPutMapper;
+import com.hav.imobiliaria.exceptions.AcessoNegadoException;
 import com.hav.imobiliaria.model.entity.Corretor;
 import com.hav.imobiliaria.model.entity.Imovel;
 import com.hav.imobiliaria.model.entity.Usuario;
 import com.hav.imobiliaria.model.enums.RoleEnum;
 import com.hav.imobiliaria.repository.UsuarioRepository;
 import com.hav.imobiliaria.repository.specs.UsuarioSpecs;
+import com.hav.imobiliaria.security.utils.SecurityUtils;
 import com.hav.imobiliaria.validator.UsuarioValidator;
 import jakarta.transaction.Transactional;
 import jakarta.validation.constraints.NotBlank;
@@ -23,6 +25,7 @@ import org.springframework.context.annotation.Lazy;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.domain.Specification;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
@@ -32,6 +35,7 @@ import java.beans.Transient;
 import java.io.IOException;
 import java.time.LocalDateTime;
 import java.util.List;
+import java.util.NoSuchElementException;
 import java.util.Optional;
 import java.util.stream.Collectors;
 
@@ -80,6 +84,10 @@ public class UsuarioService {
     }
     public Usuario salvar(UsuarioPostDTO dto, MultipartFile foto) throws IOException {
 
+        if(!dto.role().equals(RoleEnum.USUARIO) && !SecurityUtils.buscarUsuarioLogado().getRole().equals(RoleEnum.ADMINISTRADOR)){
+            throw new AcessoNegadoException();
+        }
+
         Usuario entity = instanciandoUsuarioPostDtoPorRole(dto);
 
         this.validator.validar(entity);
@@ -98,8 +106,14 @@ public class UsuarioService {
         usuarioAtualizado.setId(id);
         this.validator.validar(usuarioAtualizado);
         if(!usuarioSalvo.getRole().equals(usuarioAtualizado.getRole())) {
-            usuarioAtualizado.setId(null);
-            excluirReferenciaImovelCorretor(id);
+            if(SecurityUtils.buscarUsuarioLogado().getRole().equals(RoleEnum.ADMINISTRADOR)) {
+                usuarioAtualizado.setId(null);
+                excluirReferenciaImovelCorretor(id);
+            }
+            else{
+                throw new AcessoNegadoException();
+            }
+
         }
         Usuario usuarioJaSalvo = this.buscarPorId(id);
         if(imagemNova != null){
@@ -126,9 +140,6 @@ public class UsuarioService {
         Optional<Usuario> usuarioOptional = repository.findById(id);
         if(usuarioOptional.isPresent()){
             Usuario usuario = usuarioOptional.get();
-//            if(usuario.getFoto() != null){
-//                this.s3Service.excluirObjeto(usuario.getFoto());
-//            }
             usuario.setAtivo(false);
             usuario.setDataDelecao(LocalDateTime.now());
             this.repository.save(usuario);
@@ -217,27 +228,27 @@ public class UsuarioService {
     public List<Usuario> buscarCorretorListagem(){
         return  this.repository.findByRoleAndAtivoTrue(RoleEnum.CORRETOR);
     }
-
-    public Page<Imovel> buscarImoveisFavoritados(Long id, Pageable pageable) {
-        return repository.findImoveisFavoritadosByUsuarioId(id, pageable);
-    }
-    public  List<Long> buscarIdsImovelFavoritadoPorIdUsuario(Long idUsuario){
-        return this.repository.findIdImoveisFavoritadosByUsuarioId(idUsuario);
+    public  List<Long> buscarIdsImovelFavoritadoPorIdUsuario(){
+        Usuario usuarioLogado = SecurityUtils.buscarUsuarioLogado();
+        return this.repository.findIdImoveisFavoritadosByUsuarioId(usuarioLogado.getId());
     }
 
     @Transactional
-    public void adicionarImovelFavorito(Long idImovel, Long idUsuario) {
-        Usuario usuario = this.buscarPorId(idUsuario);
+    public void adicionarImovelFavorito(Long idImovel) {
+        Usuario usuarioAutenticado = SecurityUtils.buscarUsuarioLogado();
+        Usuario usuarioAtual = this.repository.findById(usuarioAutenticado.getId()).orElseThrow(NoSuchElementException::new);
         Imovel imovel = this.imovelService.buscarPorId(idImovel);
 
-        usuario.adicionarImovelFavorito(imovel);
-        this.repository.save(usuario);
+        usuarioAtual.adicionarImovelFavorito(imovel);
+        this.repository.save(usuarioAtual);
 
     }
     @Transactional
-    public void remocarImovelFavorito(Long idImovel, Long idUsuario) {
-        Usuario usuario = this.buscarPorId(idUsuario);
-        usuario.removerImovelFavorito(idImovel);
+    public void removerImovelFavorito(Long idImovel) {
+        Usuario usuario = SecurityUtils.buscarUsuarioLogado();
+        Usuario usuarioAtual = this.repository.findById(usuario.getId()).orElseThrow(NoSuchElementException::new);
+
+        usuarioAtual.removerImovelFavorito(idImovel);
     }
     public List<Usuario> buscarPorRole(RoleEnum role) {
         return repository.buscarPorRole(role);
